Create JWT token interface and
In JWT tokens, a claim is just something about a user, for example, a user claims that their username is 'bob'.  We then store this information in a token.   We can trust a token as it is securely signed by the API server, so if a user 'claims' their username is 'bob' then we can believe it.