Fix the 'OK' error status text [read description]
The error status text shows: 'OK'. It should be showing 'Bad Request' in case of 400 and 'Unauthorized' in case of 401. This happens because of http/2 spec and the way the Angular client defaults to displaying an OK if there is nothing in the statusText field as described here: https://github.com/angular/angular/issues/23334