EF one to many relationships [AppUser, Photos]
While EF does do the one-to-many relationships and automatically create the foreign keys in convention, it does make the user id nullable and the delete referential action to restrict not cascade. So we need to manually identify the relationships between the two tables in their entities. This is what we call "Fully-defining the relationship" between the tables.