Prepare angular app and serve it from API server
In angular.json, change outputPath to "../API/wwwroot", this is the default content folder that the Kestral server will use for static files. Our Angular application is a static file, it's just javascript files, maybe some images and that's just about it. Everything gets bundled into javascript, and javascript is static content.

Run ng build to create the static files for the angular app, they'll be placed in wwwroot folder in API.

Now when you run the API project, the angular app will run from the API, hence we can run it from port 5001 instead of 4200