Add a fallback controller
Add a fallback controller that will return the index.html [our angular app] and hence it'll start using the routing from our angular app.

Without the fallback controller, if we try to navigate to /members, we'll get nothing as the API doesn't have route for members and other angular apps.